All changes made to the code during exercises

+++ misc/objdet_tools.py
@@ -231,8 +231,12 @@ def project_detections_into_bev(bev_map, detections, configs, color=[]):
         corners_int = bev_corners.reshape(-1, 1, 2).astype(int)
         cv2.polylines(bev_map, [corners_int], True, color, 2)
 
+        #print("(corners_int.shape)", corners_int.shape)
+        #print("(corners_int)", (corners_int))
         # draw colored line to identify object front
-        corners_int = bev_corners.reshape(-1, 2)
+        corners_int = np.int_(bev_corners.reshape(-1, 2))
+        #print("(corners_int[0, 0], corners_int[0, 1])", (corners_int[0, 0], corners_int[0, 1]))
+        #print("(corners_int[3, 0], corners_int[3, 1])",(corners_int[3, 0], corners_int[3, 1]))
         cv2.line(bev_map, (corners_int[0, 0], corners_int[0, 1]), (corners_int[3, 0], corners_int[3, 1]), (255, 255, 0), 2)
